+++ commands/upload_stemcell.go
@@ -25,7 +25,7 @@ type UploadStemcell struct {
 		Stemcell   string `long:"stemcell" short:"s" required:"true" description:"path to stemcell"`
 		Force      bool   `long:"force"    short:"f"                 description:"upload stemcell even if it already exists on the target Ops Manager"`
 		Floating   bool   `long:"floating" default:"true"            description:"assigns the stemcell to all compatible products "`
-		Sha256     string `long:"sha256"                             description:"sha256 of the provided product file to be used for validation"`
+		Shasum     string `long:"shasum"                             description:"shasum of the provided product file to be used for validation"`
 	}
 }
 
@@ -66,7 +66,7 @@ func (us UploadStemcell) Execute(args []string) error {
 	}
 
 	stemcellFilename := us.Options.Stemcell
-	if us.Options.Sha256 != "" {
+	if us.Options.Shasum != "" {
 		shaValidator := validator.NewSHA256Calculator()
 		shasum, err := shaValidator.Checksum(stemcellFilename)
 
@@ -74,8 +74,8 @@ func (us UploadStemcell) Execute(args []string) error {
 			return err
 		}
 
-		if shasum != us.Options.Sha256 {
-			return fmt.Errorf("expected shasum %s does not match file shasum %s", us.Options.Sha256, shasum)
+		if shasum != us.Options.Shasum {
+			return fmt.Errorf("expected shasum %s does not match file shasum %s", us.Options.Shasum, shasum)
 		}
 
 		us.logger.Printf("expected shasum matches stemcell shasum.")