+++ commands/upload_stemcell_test.go
@@ -246,7 +246,7 @@ var _ = Describe("UploadStemcell", func() {
 		})
 	})
 
-	Context("when the --sha256 flag is defined", func() {
+	Context("when the --shasum flag is defined", func() {
 		It("proceeds normally when the sha sums match", func() {
 			file, err := ioutil.TempFile("", "test-file.tgz")
 			Expect(err).ToNot(HaveOccurred())
@@ -269,7 +269,7 @@ var _ = Describe("UploadStemcell", func() {
 			command := commands.NewUploadStemcell(multipart, fakeService, logger)
 			err = command.Execute([]string{
 				"--stemcell", file.Name(),
-				"--sha256", "2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44",
+				"--shasum", "2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44",
 			})
 			Expect(err).NotTo(HaveOccurred())
 			format, v := logger.PrintfArgsForCall(0)
@@ -288,7 +288,7 @@ var _ = Describe("UploadStemcell", func() {
 			command := commands.NewUploadStemcell(multipart, fakeService, logger)
 			err = command.Execute([]string{
 				"--stemcell", file.Name(),
-				"--sha256", "not-the-correct-shasum",
+				"--shasum", "not-the-correct-shasum",
 			})
 			Expect(err).To(HaveOccurred())
 			Expect(err).To(MatchError("expected shasum not-the-correct-shasum does not match file shasum 2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44"))
@@ -297,7 +297,7 @@ var _ = Describe("UploadStemcell", func() {
 			command := commands.NewUploadStemcell(multipart, fakeService, logger)
 			err := command.Execute([]string{
 				"--stemcell", "/path/to/testing.tgz",
-				"--sha256", "2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44",
+				"--shasum", "2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44",
 			})
 			Expect(err).To(HaveOccurred())
 			Expect(err).To(MatchError("open /path/to/testing.tgz: no such file or directory"))
@@ -356,7 +356,7 @@ var _ = Describe("UploadStemcell", func() {
 		BeforeEach(func() {
 			var err error
 			configContent := `
-sha256: 2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44
+shasum: 2815ab9694a4a2cfd59424a734833010e143a0b2db20be3741507f177f289f44
 `
 			configFile, err = ioutil.TempFile("", "")
 			Expect(err).NotTo(HaveOccurred())